Refactoring to extract DocProperties Reader from Gnumeric Reader
